Update patch set 1
Patch Set 1:

> > The concern about a mutex/cond variable for each I/O is that on
 > FreeBSD, creation of such is not as trivial as it is in Linux.
 > 
 > Ah, FreeBSD cannot just do struct async_..._data data = { .mutex = 
 > PTHREAD_MUTEX_INITIALIZER, ... }; ?
 > I didn't know that, would explain...

It can do that, but some things that have been said in commit messages for FreeBSD compatibility patches seem to me that there's a bit more to mutex and condition variable creation and destruction for FreeBSD  than for Linux, I'd love to understand that. If it's really almost trivial to create/destroy, then the mutex and condition variable could live in fsal_read and fsal_write and be much safer.

 > > Note that the current structure is mutex/cond variable per
 > thread. Each thread can only have one synchronous I/O in
 > progress...
 > 
 > I agree it shouldn't happen with the current code, it's just you're
 > leaving it possible through the API so someone might do it
 > eventually - I don't like that :)
 > 
 > > It is a bit annoying that the caller has to provide them, but it
 > seemed like the easiest way to not have a single global mutex and
 > avoid creating and destroying them all the time.
 > 
 > Yeah if freebsd needs to call pthread_mutex_init/destroy all the
 > time I agree it's annoying... I guess I'm still on the edge, but in
 > this case the balance does tip over to passing the argument.. Or
 > maybe we could get it through tls? is op_ctx still used? (I never
 > liked it though, so, overall actually still thinking the way you
 > did is probably the most straightforward/best overall)

I thought about op_ctx, but that raised issues of making the lifetime clear. The way I implemented it feels like it makes the lifetime much more clear. 

 > All that being said, I'd just change the pthread_cond_signal to a
 > broadcast to be on the safe side if someone ever misuses this and
 > reuses the same mutex/cond somewhere by mistake and keep the rest
 > as is.

That may be a good idea. I suppose if there's only a single thread waiting, it won't be any more costly than a signal.

Patch-set: 1
Reviewer: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>