Update patch set 1
Patch Set 1:

>  > Ah, FreeBSD cannot just do struct async_..._data data = { .mutex = 
>  > PTHREAD_MUTEX_INITIALIZER, ... }; ?
>  > I didn't know that, would explain...
> 
> It can do that, but some things that have been said in commit messages for FreeBSD compatibility patches seem to me that there's a bit more to mutex and condition variable creation and destruction for FreeBSD  than for Linux, I'd love to understand that. If it's really almost trivial to create/destroy, then the mutex and condition variable could live in fsal_read and fsal_write and be much safer.

Would be great if we could clarify that... Who are our current FreeBSD folks? Added Fatih to the Cc list, if you know more about pthread mutex/cond initialisation, please speak up :)


>  > All that being said, I'd just change the pthread_cond_signal to a
>  > broadcast to be on the safe side if someone ever misuses this and
>  > reuses the same mutex/cond somewhere by mistake and keep the rest
>  > as is.
> 
> That may be a good idea. I suppose if there's only a single thread waiting, it won't be any more costly than a signal.

The userspace implementation I looked at (glibc, nptl) is really mostly the same, the only difference is it calls futex_wake with INT_MAX instead of 1 -- and the linux kernel implementation just breaks when that count is reached so in the case of one waiter the cost is strictly the same.

Patch-set: 1
CC: Fatih Acar <1014381@d5d70762-12d0-45a1-890d-524b12d3f735>