support: add a gsh_refstr type and functions
Create a new structure that consists of an atomic refcount value and a
flexarray that is intended to hold a string. When the refcount reaches
zero, then the object will be freed.

Unfortunately, the liburcu version currently in epel7 doesn't have
urcu_ref_get_unless_zero, so a reimplementation is added for when
it isn't present.

Change-Id: I742e1fff3e47646e64fec235fad461b206924458
Signed-off-by: Jeff Layton <jlayton@redhat.com>