Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1:
> 
> 1. let us say we want to move object A to directory C (from B).
> 2. rename is done in FSAL, so A is now visible in C
> 3. Another move of A from C to D. Let us say this also calls FSAL rename, so A is now visible in directory D.
> 
> These two operations now try to set the parent pointer for A at the same time. There is no ordering, so it could be set to D first and then to C. Finally it ends up pointing to C which is incorrect (should be D).

Okay, I can see that.  I'm torn between this, and changing it to mdc_get_parent().  On the one hand, it's an extra call to the sub-FSAL, adding latency to rename.  On the other, it will need to be done next time we do a lookupp(), which will add latency to that.  I guess it doesn't matter too much, so I'll operate on the assumption lookupp() is relatively rare.

Patch-set: 1
Reviewer: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1