Update patch set 1
Patch Set 1:

> AFAIK the only difference between FreeBSD's libpthread implementation and glibc's is about object allocation. pthread objects are dynamically allocated on FreeBSD: we must destroy objects otherwise leak happen. Using PTHREAD_MUTEX_INITIALIZER works as the lib is automatically allocating required stuff on first call (cf. https://github.com/freebsd/freebsd/blob/master/lib/libthr/thread/thr_mutex.c#L584).

Thanks for checking!
So ultimately it's a matter of style (keeping the mutex/cond local to the fsal_read/write functions) vs. the cost of one allocation/free everytime we do an IO...
For me, if we keep small I/Os (e.g. <=64k?) synchronous I'd say the overhead cost is acceptable (negligible compared to a large-ish IO); but if we use the same async mechanism for small IOs then we probably should keep it as you've done in this version.

Ultimately I'm fine with either version, just not giving +1 here because if we keep the current code I'd really prefer the signal to be a broadcast :)

Patch-set: 1