Update patch set 1
Patch Set 1:

> I don't think it would be good to change things so multiple threads would actually use the same condition variable. The problem there is then that you have a thundering herd wake up. The design is that each I/O has a boolean that indicates it is complete. So exactly one I/O has completed, and there should be exactly one thread waiting for that I/O, so pthread_cond_signal should be perfect.
> 
> Anything different implies the caller actually supports some degree of asynchronous I/O and therefore will not use fsal_read/fsal_write which is intended to make the async I/O look like a blocking read or write call (i.e. it doesn't return until the I/O is complete). Yea, we make the caller pass in the mutex and condition variable, but that's to avoid creating and destroying them for each I/O, not for them to be shared.

Hmm, yes and no. If you just change the signal to a broadcast we will effectively support sharing the lock, even if it's less efficient - we still have the boolean to go back to sleep. Yes it's less efficient and I wouldn't recommend anyone to use it, but it'll work consistently.

If we don't and someone uses the same cond by mistake, it will appear to work in simple cases but in higher load will just start hanging with no apparent reason.
The argument "one thread needs waking up to match the boolean update" is correct but in the (incorrect) case multiple threads are waiting you don't know which one will be woken up, and if the wrong one gets woken up it'll just go back to sleep, then when the next one gets completed it'll wake either thread up, and the last thread will be waiting for that cond forever...

So long story short, it's not about boasting support for sharing cond/mutex, but just about being more robust in case some later user happens to do it by mistake.
It's less costly to make that a broadcast than to try to detect people using it badly; the one thing I want to avoid is something that eventually screws up once in a blue moon and gets stuck for no apparent reason only under very specific conditions that'll be impossible to reproduce...

Patch-set: 1