Update patch set 2
Patch Set 2:

I actually implemented this at one point when working on readdir_plus, but it turned out to not be necessary for correctness.

This is supposed to be handled by READAHEAD.  It isn't in gluster at the moment, because gluster has no way to know when the set of FDs returned by a readdir_plus call is done, so it keeps going and enumerates everything.  I wrote up a quick PoC a while ago, and Jiffin submitted it, but it wasn't accepted:

https://review.gluster.org/#/c/glusterfs/+/20038/

I'd prefer that we not add dirfd caching if we can avoid it; it adds a fair amount of complexity, and a chance to leak FDs.

Patch-set: 2
CC: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>