Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > I think we need more logic in state_deleg_conflict to check if the
>  > file has been delegated to an unique client and the operation is
>  > done by the same client. Otherwise it would add an useless delay to
>  > these operations...
> even for same client if the operation is for unlink,rename,link,chmod, truncate server should recall AFAIK

According to RFC5661:

"If the object whose attributes are being changed has a file
   delegation that is held by a client other than the one doing the
   SETATTR, the delegation(s) must be recalled"

" If the object being renamed has file delegations held by clients
   other than the one doing the RENAME, the delegations MUST be
   recalled, and the operation cannot proceed until each such delegation
   is returned or revoked."

It seems pretty obvious that we should not delay if the delegation is uniquely held, it may have significant performance impact on some workloads. We should at least check this case in state_deleg_conflict.

Patch-set: 1