Update patch set 2
Patch Set 2:

> > > This is supposed to be handled by READAHEAD.  It isn't in gluster at the moment, because gluster has no way to know when the set of FDs returned by a readdir_plus call is done, so it keeps going and enumerates everything.  I wrote up a quick PoC a while ago, and Jiffin submitted it, but it wasn't accepted:
> > > 
> > > https://review.gluster.org/#/c/glusterfs/+/20038/
> > 
> > I have reviewed it. It's a great idea.
> 
> It'd be great if someone could take that over and shepherd it through.  I don't have time right now.

I'd like to pick it up and finish it base on your and Kaleb's pull requests if you don't mind.

https://review.gluster.org/#/c/glusterfs/+/20038/
https://review.gerrithub.io/c/ffilz/nfs-ganesha/+/428230

But for the ganesha part, you comment as,
"To fix this, open dir handles when the first dirent is read from the FSAL, and leave them open until either 1) the readdir completes on that director; or 2) the LRU demotes the handle and closes it."

#1 is finish in your patch, but #2 is not, right?
I think my patch part #2.
So, I want to merge your patch with my patch.

Patch-set: 2
Reviewer: Kinglong Mee <1011834@d5d70762-12d0-45a1-890d-524b12d3f735>