Update patch set 1
Patch Set 1:

Hmmm... I'm not sure I agree...  Failure of this is not propagated up, so a caller cannot know it failed.  The only indication is this log message, and putting it at Event makes it seem like a normal occurrence which I don't think it is.  I'd say it should be at least Warn, if not Crit.

However, maybe I'm misunderstanding the cause?  What causes a mountpoint to not fail stat(), and how often does it occur?  It seems to me stat() should always succeed on a mountpoint unless there's major error of some kind.

Patch-set: 1
CC: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>