cmake: work around lttng/urcu/_LGPL_SOURCE incompatibility
When _LGPL_SOURCE is defined, the lttng and urcu code will turn some
library functions into macros or static inlines. Unfortunately, when
lttng does this, it always selects the "bulletproof" (bp) flavor of
urcu.

With the new usage of urcu in ganesha this is problematic as we may
not have liburcu-bp linked in.

Undefine _LGPL_SOURCE for now until we have a proper fix.

Change-Id: I59c900a7af39920261b12f04e0ec3c2fafd3a8da
Signed-off-by: Jeff Layton <jlayton@redhat.com>