Update patch set 1
Patch Set 1:

I'm not thrilled with this patch, but we do need something like it. We could also consider undefining this in a smaller scope, only when the lttng headers are being included.

Another potential fix would be to switch back to forcing urcu-bp. It's not as performant in general but is probably good enough for our current usage.

Patch-set: 1
Reviewer: Jeff Layton <1008909@d5d70762-12d0-45a1-890d-524b12d3f735>