Update patch set 2
Patch Set 2: Code-Review-1

Why this change?  deledrecall_one() says in it's comment that the lock must be taken.  I think the first change was correct.

If this is correct and the first change is not, then the comment needs to change.

Patch-set: 2
Reviewer: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1