Remove unused posix mount details when adding export if duplicate is found
Ganesha maintains a list of mount points on the host to map the exports to
a particular mount point. If the mount point is removed and remounted, the
device id (st_dev) is supposed to change.

A side effect of this mechanism is that, if a mount is removed (unmounted)
from the system, ganesha still holds onto its details. If at a later point
another mount is added which gets the same device id (stat.st_dev) as the
old mount, ganesha detects the new mount as a duplicate file system.

This change makes sure that, if the device id is reused and ganesha has old
mount details in memory, we remove the old mount details if it is no longer
mounted and exported by ganesha. If the device is reused, the old mount is
definitely unmounted, so we just have to check that it is not being
exported by ganesha to safely remove it.

Change-Id: I97d63f5bc21e29bc3ff7f73bf78a7208a3a41f6e
Signed-off-by: Sriram Patil <sriramp@vmware.com>