Update patch set 1
Patch Set 1:

(2 comments)

> (2 comments)
 > 
 > I'm really not sure this is the right solution.
 > 
 > I feel like we're starting to build hacks on top of what admittedly
 > was originally somewhat of a hack.
 > 
 > The problem we face is that there are various things (that I don't
 > fully understand) that make a given path show up multiple times in
 > the filesystem list. Also, there are various special types of
 > filesystem that don't have a unique device id.
 > 
 > And then we throw in unmounting and mounting filesystems while
 > Ganesha is running...
 > 
 > We may need to start back from scratch and figure out a better way
 > to find filesystems.

Actually, device ids should be unique if all the mounts are live. If we unmount a filesystem and then mount some other file system, the new mount can end up getting the previous device id, because it is now reused. This is valid, because device ids are still unique on the system. We have a problem in this case because ganesha does not clear the mount points from the list of mounts.

So, when we get a duplicate fsid/device id we can safely assume that the old mount was unmounted and discard it if it is not exported.

There is one problem with this fix though, it can happen that the parent mount is not exported but one of the children can be exported and we are still removing it from the list of mounts. This can cause issues.

Patch-set: 1
Reviewer: Sriram Patil <1010034@d5d70762-12d0-45a1-890d-524b12d3f735>