Update patch set 5
Patch Set 5:

> Posted a couple of comments on github.
 > Independantly, I think the pullup should be merged with the
 > previous commit so ganesha always compile. ntirpc pullup with API
 > change make sense to have in the same commit as the API change on
 > ganesha side

I'll fix the issue in ntirpc.

This patch will be changed to a pullup of the actual ntirpc, and squashed with the previous patch, I wanted to keep this separate to make it easy to remember that swizzle needed to be done before merge.

Patch-set: 5
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>