Modify strmaxcpy definition to be more graceful!
Currently, strmaxcpy() doesn't change the dest buffer if the src buffer
couldn't be copied in full. This implies that the callers of strmaxcpy
should always check for the return code unless we known for sure that
the src length is small in which case we can use strcpy itself.

There are callers of strmaxcpy that don't check the return value because
they use the dest string for logging purposes only and are lazy.
Ideally, strlcpy is best suited for these callers but making strmaxcpy
behave like strlcpy (copying a truncated string) doesn't seem to have
any downsides. The current implementation of strmaxcpy also forces us to
scan the src string twice.

Change-Id: I06160ea14155a9000bdba4e331a5aa361921d255
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>