CB_GETATTR implementation
This patch adds CB_GETATTR op support. This callback op is used
to fetch filesize and change attributes from the client holding write
delegation on a file. This is needed to provide accurate attribute
information to other clients doing GETATTR on the same file.
Till we get response for this callback, NFS4ERR_DELAY error is sent
to those other clients.

Change-Id: I402d80aad591d9e2ae51e3dd9c279b662dabe793
Signed-off-by: Soumya Koduri <skoduri@redhat.com>