Update patch set 3
Patch Set 3:

I think a lot of these should use strmaxcpy and check for error.

For those that are copying a constant string, it would be nice if we could flag at compile time that the constant string is longer than the buffer...

Honestly I'd also like to better distinguish the cases where it's ok to truncate and continue on.

Patch-set: 3
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>