Update patch set 4
Patch Set 4:

> (1 comment)
 > 
 > > Patch Set 3:
 > >
 > > I think a lot of these should use strmaxcpy and check for error.
 > >
 > > For those that are copying a constant string, it would be nice if
 > we could flag at compile time that the constant string is longer
 > than the buffer...
 > >
 > > Honestly I'd also like to better distinguish the cases where it's
 > ok to truncate and continue on.
 > 
 > I don't see a difference between using strmaxcpy vs strlcpy. They
 > both provide return values to check if the dest buffer is enough or
 > not. strlcpy is actually better in this regard as you can allocate
 > a bigger buffer if needed as part of recovery.

So what might be a good division:

Use strmaxcpy where there is a problem if we overflow the buffer and we can't re-allocate.

Use (void) strlcpy if truncation is ok. We should also verify that none of these use UTF8
strings and if so, use display buffer instead (or we could pull the guts of _display_complete_overflow so that we can use the same functionality with strlcpy).

Use ret = strlcpy() in places where we can re-allocate the buffer (I don't know if we actually have any of these).

Consider using strcpy for cases where the source is a fixed length and we know the destination will be big enough.

If we're going through the effort of looking at each place a string is copied let's make them all right...

Patch-set: 4
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>