RFC: [do not merge] (lib)ganesha_nfsd as a DSO
Build ganesha as a DSO, i.e. libganesha_nfsd.so.

Take No. Two-dot-one, addressing comments from the first go-around.

The ganesha.nfsd daemon consists of just nfs_main:main() and is
linked with the DSO.

Then all the FSALs (except FSAL_PSEUDO and FSAL_MDCACHE*) are linked
with libganesha_nfsd resolution of functions in ganesha.nfsd itself.

This will be required if, e.g., Fedora, decides to require default
linking with '-z defs'. They tried to do it for F29 but retreated
when they found that too many applications broke. In the end this may
all be just an exercise in "truth and beauty."

*FSAL_PSEUDO and FSAL_MDCACHE are special; they are included in the
DSO so do not need to be (and can't be) linked with the DSO.

A few other things that could be done are
 1) link libganesha_nfsd.so with a .sym file to explicitly control which
    symbols are exposed or visible to the outside world.
 2) build with -flto (Link Time Optimization)
    Annecdotes claim a 5-10% speed improvement is typical, and some
    applications have seen as much as 45-50% boost. Of course we have to
    do the experiment to see what we really achieve.
 3) don't link ganesha.nfsd with -bdynamic as it's no longer needed

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Change-Id: Iab35f1d9bb1da02685bdaadea2a838d2ec5ee6ec