Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> this is fine of course, but in a code path this small, it doesn't simplify very much

Yes. The main reason for it is that a later patch needs to be able to set up a RADOS client from another codepath, so we do need this in a helper.

I probably should have made that clear in the description...

Patch-set: 1
Reviewer: Gerrit User 1008909 <1008909@d5d70762-12d0-45a1-890d-524b12d3f735>