Drop dirent ref when releasing dirent
During a readdir, the dirents in a chunk take a ref on the entry they
point to, so that the entry can't be reaped until the chunk has been
processed.  Make sure that, if a dirent is released while it still has a
ref, we drop the ref.  Otherwise, under heavy create/delete/readdir
load, we can leak refs to entries.

Change-Id: Id759bb40f33c92867c1fda93f9449433f686d4d5
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>