Update patch set 4
Patch Set 4:

> I don't actually use this file, I just cleaned it up for others. I tried to use this file and/or what's on the wiki. That didn't work, so I went through the code and figured out what was needed.

Still, you've basically reverted to the state it was in six months ago, so I assume you needed to add back the PROXY { Remote_Server {}} block that I still think is no longer used? I don't have time to test this today but will give it another run tomorrow to check

As for the wiki, it's pointing to the manpage directory of the next branch as far as I can see, so it also looks up to date to me. I assume you are running off next and not master?

Patch-set: 4