Update patch set 1
Patch Set 1: Code-Review-1

Ok, this removed UTF8_SCAN_CKUTF8 which checks for malformed UTF8, which is why all the pynfs UTF8 tests are failing.

What special characters are you tripping over?

I wonder if what is happening is you have non-canonical UTF8 where a character that can be represented in fewer bytes is being represented in more (for example 'A' can be encoded as 0x41 or 0xC1 0x81). This is overlong encoding and is not supposed to be used (except 0xC0 0x80 may be allowed to encode the NUL character in a string terminated by 0x00).

Patch-set: 1
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1