Update patch set 1
Patch Set 1: Code-Review-2

Consensus seems to be that we don't want to make an exception to checking for proper UTF-8. There may be a consideration to allow the surrogate pairs that Windows uses, but it appears that the test case that the end user is trying that prompted this change was to submit strings with every byte code from 0x01 to 0xff excluding '/' (not sure if the end user test tried to present the "." or ".." strings, but they definitely need to be excluded also). The problem is that "random" byte codes from 0x01 to 0xff will result in partial UTF-8 characters which may be problems in real situations. The fact that EXT4 allows such strings should not be reason to ignore the NFS v4.x RFCs.

If a clear use case can be presented, we could entertain a configurable alternate string validation that would allow non-UTF-8 strings (but might then do other validation checks), however, such a use case is probably getting into using code pages and stuff that Unicode (and the UTF-8 expression thereof) was intended to make less problematic.

Patch-set: 1
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-2