Update patch set 3
Patch Set 3:

> Do we update change time for added directory? That would have the
 > same issue.

The change attribute is used for invalidating the child dentry attributes. On creation, a completely new dentry is created so it should not ideally invalidate cached attributes of all the other dentries under that directory.

Only in case of removal, we need to make sure that the older entries are removed from the cache otherwise a new entry with same name can result in ESTALE.

Updating change time on mkdir/create makes sense. But it will result in invalidation of of all the dentries under that directory.

Patch-set: 3
Reviewer: Gerrit User 1010034 <1010034@d5d70762-12d0-45a1-890d-524b12d3f735>