Update patch set 1
Patch Set 1: Code-Review-1

This, in fact, won't work.  We can't depend on the order of the list at all, since chunks can be removed from the middle of it, and we can start a readdir with a new chunk in the middle.  I'm working on a proper fix for this.

Patch-set: 1
Reviewer: Gerrit User 1004087 <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1