Update patch set 1
Patch Set 1:

I'm not sure this is quite the final version, but let's start reviewing it.

Sriram - this includes your patch, however, I will merge your patch separately since it needs back porting to V2.7.2 and this patch can't be back ported since it changes the FSAL API.

Malahal - while working on this, I noticed that FSAL_GPFS computes change differently that the other FSALs, we might want to change that either as part of this patch or as a separate patch.

Daniel - Working on this I notice that FSAL_MEM has a bit of wonkiness on change attribute, it could actually use a pure monotonic counter for change, on the other hand, it is also fine for it to assume that if mtime or ctime is updated to current time, then just using that for the new change is reasonable. There is one place it makes an increment to change, I know mdcache increments change in certain cases, but this may be different?

Patch-set: 1
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>