Bump the lease timer when Clientid is reused after lease expires
If a RENEW request comes after the lease interval time, there can
be a loop of NFS4ERR_EXPIRED errors with RENEW, SETCLIENTID,
SETCLIENTID_CONFIRM and other requests.

This happens due to the existing client record being reused in the
small time window till the reaper is run.

Reset the cid_last_renew when we confirm the client so that it
gets the correct lease interval to renew its lease with the
next request.

Change-Id: I6b4c5c39aa804b3bad60d584daa1d56dd62c6d8b
Signed-off-by: Trishali Nayar <ntrishal@in.ibm.com>