rpm: fix selinux macros within mock
We build the SELinux policy with /usr/share/selinux/devel/Makefile, so
we must BuildRequire the selinux-policy-devel package.

Also, we must tell rpmbuild to ignore the selinux_requires macro if
it is undefined, because rpmbuild evaluates the spec file very early in
the build process, before mock installs the BuildRequires packages.
Once mock installs selinux-policy, the macros will be available, but we
need to gracefully handle the case where %selinux_requires macro is not
yet avaialable.

Change-Id: I7ceeef7c0ae5652fa11963a807cded661f174d4f
Signed-off-by: Ken Dreyer <kdreyer@redhat.com>