Drop dirent ref before mdcache key delete in avl_dirent_set_deleted
Was hitting assert(FSAL_IS_SUCCESS(status)); in mdcache_avl_remove.
We have taken ref on the dirent which was supposed to be dropped by
mdcache_readdir_chunked. But before that, in between readdir RPCs,
an unlink/rename could come in which calls mdcache_dirent_remove.
Currently we are calling mdcache_key_delete in avl_dirent_set_deleted
without checking for ref.

Signed-off-by: Ashish Sangwan <ashishsangwan2@gmail.com>
Change-Id: If88de31c501396dc513459821af41c3e7ee47438