Update patch set 2
Patch Set 2:

Is there any possibility this causes other legitimate permission checks to not kick in properly?

The ideal is to check if a read only file is being opened for write by the owner. The other check to make is for an execute only file being opened for read (which this change also happens to allow).

Patch-set: 2
CC: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>