Update patch set 2
Patch Set 2:

> I have write a short code tested open read only file by O_RDWR, and open failed. 
> I think in other case of open will not run into the func glusterfs_open2's label "open" code.

That could be because the permission check would have got done in upper layers. In glusterfs_open2(), we directly goto label "open" if the createmode is set to FSAL_NO_CREATE. But I see that in that case we set "caller_perm_check" to true to let the caller do permission checking post opening the file [note: which is not required as we set and restore caller creds while opening fd].

So this change may seem working now but will not work if the ganesha process is started by a non-root user I think.

Patch-set: 2