Update patch set 2
Patch Set 2:

> > Patch Set 2:
 > >
 > > > Patch Set 2:
 > > >
 > > > (1 comment)
 > >
 > > May be glusterfs caused the bug, but I am not familiar with
 > glusterfs code. I am a newbie to glusterfs and ganesha.
 > > For this problem, I analysis when copy read only file error, the
 > glusterfs server log. In log, I see ganesha first create the file
 > and then open the file, I think when wirte to gluster file system,
 > ganesha doesn't use the file handle which returned by create func
 > but he handle which returned by open function.
 > 
 > I understood the problem.
 > 
 > glusterfs allows RDWR permission to the owner for the fd returned
 > through CREATE call but not through OPEN, which I think is the
 > right thing to do as other filesystems do the same. The issue now
 > is that as you mentioned, nfs-ganesha doesnt use the handle
 > returned via glfs_h_creat but has to open the file again to fetch
 > fd for write operations.
 > 
 > I am not sure how kNFS addresses this problem. But certainly the
 > fix posted here is not right and exposes a security caveat where in
 > any user can open a RO file with RDWR access (if somehow get past
 > permission checking done in upper layers).
 > 
 > Maybe we can modify glfs_h_creat or define new api in libgfapi to
 > return both handle and fd as part of single creation call. That IMO
 > should fix this issue and also avoids unnecessary second OPEN call
 > for I/O operations. Thoughts?

That would actually be the ideal solution. Having an atomic create and open really solves the issue, though there may still be an issue on state recovery. In the case of state recovery, we have to open a file where as far as the client process is concerned, is still open, so a reclaim open needs to go through as long as it's from the owner of the file.

Stateless I/O (NFS v3 and NFS v4 using the special stateids) also needs to be able to access a file where the permissions do not match as long as the user if the owner of the file. The upper layers of Ganesha check this (and note that a client side open DOES make an ACCESS call, so the client won't allow a process to open a read only file as read/write, but would allow an open/create of a read only file as read/write. This combined with only allowing the owner to override the permissions prevents any actual security hole.

The other issue is that the server must allow clients to read execute only files. NFS doesn't have an EXECUTE operation, so must issue READ operations to read in binary files. Again, the client will have done an ACCESS check and would prevent a process from opening an execute only file for read, reading an execute only file will only be allowed to a process that is specifically allowed to execute files (so bash for example can read execute only files).

kNFS and the kernel does have logic to allow the owner override. I'm not sure it works everywhere.

Patch-set: 2
Reviewer: Gerrit User 1003571 <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>