ganesha_mgr gets a new option to show the filesystem cache
The "show fscache" option is added to the ganesha_mgr utility.
It will show the major and minor device id's of the mounted
filesystems.

The "show" subcommand can also be used to display exports and
clients via "show exports" and "show clients" repectively.

The older interfaces of show_exports and show_client are still
retained.

Change-Id: If18b7a45918079ac3ccba1a2a1c5855fb1d17b53
Signed-off-by: Trishali Nayar <ntrishal@in.ibm.com>