SAL: switch to just tracking long-form clientid strings
We currently store a combination of the long-form clientid, its length
and the client's address. Storing the address here means that we can't
deal with trunking in the future. We should just be storing the long
form clientid and using that for comparison.

Change the code to just store the long-form clientid blob.

Just changing that however would break recovery after an upgrade.
Luckily all the recovery backends generated the string the same way.

Add a routine to generate a legacy string for this. If we don't match
an existing client with the new style strings, generate a legacy form
cid_recov_tag and see if it matches any of them.

Change-Id: Ibc600b95c625910ec8c8a0e0ac80a81ea67cc570
Signed-off-by: Jeff Layton <jlayton@redhat.com>