rados_kv: don't track the client's address in the recovery database
We currently store a combination of the long-form clientid, its length
and the client's address. Storing the address here means that we can't
deal with trunking in the future. We should just be storing the long
form clientid and using that for comparison.

Unfortunately, this change may break reboot recovery so we will probably
want to only add this patch to the next major release.

Change-Id: Ibc600b95c625910ec8c8a0e0ac80a81ea67cc570
Signed-off-by: Jeff Layton <jlayton@redhat.com>